Extract runner lane CLI commands
